scraper/requirements.txt
beautifulsoup4
mysql-connector-python
flask
requests
pika
